Resume_Pipeline/requirements.txt
openai
PyMuPDF
pdfplumber
pdf2image
pytesseract
ocrmypdf
requests
python-dotenv
fastapi
uvicorn[standard]
python-multipart
python-docx